+++ src/main/java/com/hhpl/concertreserve/application/facade/ConcertFacade.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application;
+package com.hhpl.concertreserve.application.facade;
 
 import com.hhpl.concertreserve.domain.concert.*;
 import com.hhpl.concertreserve.domain.concert.model.Concert;
