+++ src/main/java/com/hhpl/concertreserve/application/facade/PaymentFacade.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application;
+package com.hhpl.concertreserve.application.facade;
 
 import com.hhpl.concertreserve.domain.concert.ConcertService;
 import com.hhpl.concertreserve.domain.concert.model.Reservation;
