+++ src/main/java/com/hhpl/concertreserve/application/facade/WaitingQueueFacade.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.application;
+package com.hhpl.concertreserve.application.facade;
 
 import com.hhpl.concertreserve.domain.waitingqueue.model.WaitingQueue;
 import com.hhpl.concertreserve.domain.waitingqueue.model.WaitingQueueInfo;
